# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OSS Continuing Education Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}/balance:
    get:
      operationId: getUserBalance
      tags: [users]
      summary: Get a user's current con-ed balance for this year
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balance info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceInfo"

  # ─── Object Storage ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    BalanceInfo:
      type: object
      required: [userId, annualAllocation, usedAmount, remainingAmount, year, isProrated]
      properties:
        userId:
          type: integer
        annualAllocation:
          type: number
          format: float
        usedAmount:
          type: number
          format: float
        remainingAmount:
          type: number
          format: float
        pendingAmount:
          type: number
          format: float
        year:
          type: integer
        isProrated:
          type: boolean
        hireMonth:
          type: ["integer", "null"]

    # ─── Con-Ed Request ────────────────────────────────────────────────────────
...