# lib/api-zod/src/generated/types/balanceInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OSS Continuing Education Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface BalanceInfo {
  userId: number;
  annualAllocation: number;
  usedAmount: number;
  remainingAmount: number;
  pendingAmount?: number;
  year: number;
  isProrated: boolean;
  /** @nullable */
  hireMonth?: number | null;
}